Expendature_System/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "fastmcp>=2.12.5",
    "google-genai>=1.46.0",
    "mysqlclient",
    "openai>=2.6.0",
    "pymysql>=1.1.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
